adk-evaluation/pyproject.toml
[project]
name = "adk-evaluation-practice"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.23.0",
]

[dependency-groups]
dev = [
    "google-adk[eval]>=1.23.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest]
testpaths = ["tests"]
filterwarnings = [
    'ignore:\[EXPERIMENTAL\]:UserWarning'
]
